requirements.txt
six
dwave_networkx>=0.6.0
pysmt
penaltymodel